Safari-609.1.11.1/LayoutTests/http/tests/security/mixedContent/insecure-image-with-securecookie-expected.txt

CONSOLE MESSAGE: line 3: secureCookie=yes
CONSOLE MESSAGE: line 5: The page at https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-image-secure-cookie.html was allowed to display insecure content from http://127.0.0.1:8080/security/resources/compass.jpg.

This test opens a window that tries to read a secure cookie and then load an insecure image. This should not block loading of the insecure image, even though a secure cookie was accessed on this page, because insecure images can only affect pixels on the screen, not expose private data.
